Rename config_layer_provider.hpp
