+++ Porytiles2/include/porytiles2/infra/config/default_provider.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "porytiles2/infra/config/config_layer_provider.hpp"
+#include "porytiles2/infra/config/config_provider.hpp"
 
 namespace porytiles2 {
